PUB-1996 - Updated chart + dependencies (#233)
* PUB-1996 - Updated chart + dependencies

* PUB-1996 - Updated suppressions

* PUB-1996 - Updated gradle

* PUB-1996 - Updated sonar

* PUB-1996 - Updated subscription management

* PUB-1996 Updated structure

* PUB-1996 - Further updates to build gradle

* PUB-1996 - Updated to functional

* PUB-1996 - Updated functional tests

* PUB-1996 - Fix sonar + bug in test

* PUB-1996 - Fixed unused import

* PUB-1996 - Updated data models
