+++ README.md
@@ -189,7 +189,7 @@ The service can also be adapted using the yaml files found in the following loca
 - [src/main/resources/application.yaml](./src/main/resources/application.yaml) for changes to the behaviour of the service itself.
 - [src/main/resources/application-dev.yaml](./src/main/resources/application-dev.yaml) for changes to the behaviour of the service when running locally.
 - [src/test/resources/application-test.yaml](./src/test/resources/application-test.yaml) for changes to other test types (e.g. unit tests).
-- [src/integrationTest/resources/application-integration.yaml](./src/integrationTest/resources/application-integration.yaml) for changes to the application when its running functional tests.
+- [src/integrationTest/resources/application-functional.yaml](./src/functionalTest/resources/application-functional.yaml) for changes to the application when its running functional tests.
 
 ## API Documentation
 Our full API specification can be found within our Swagger-UI page.
