+++ build.gradle
@@ -4,13 +4,13 @@ plugins {
   id 'pmd'
   id 'jacoco'
   id 'io.spring.dependency-management' version '1.1.0'
-  id 'org.springframework.boot' version '2.7.11'
+  id 'org.springframework.boot' version '2.7.12'
   id 'org.owasp.dependencycheck' version '8.2.1'
-  id 'org.sonarqube' version '4.0.0.2929'
+  id 'org.sonarqube' version '4.2.0.3129'
   id 'com.github.ben-manes.versions' version '0.46.0'
   id 'io.freefair.lombok' version '8.0.1'
   id 'org.jetbrains.kotlin.jvm' version '1.8.21'
-  id 'org.flywaydb.flyway' version '9.17.0'
+  id 'org.flywaydb.flyway' version '9.19.1'
 }
 
 apply plugin: 'org.owasp.dependencycheck'
@@ -31,18 +31,9 @@ sourceSets {
     java {
       compileClasspath += main.output
       runtimeClasspath += main.output
-      srcDir file('src/integrationTest/java')
+      srcDir file('src/functionalTest/java')
     }
-    resources.srcDir file('src/integrationTest/resources')
-  }
-
-  integrationTest {
-    java {
-      compileClasspath += main.output
-      runtimeClasspath += main.output
-      srcDir file('src/integrationTest/java')
-    }
-    resources.srcDir file('src/integrationTest/resources')
+    resources.srcDir file('src/functionalTest/resources')
   }
 
   smokeTest {
@@ -58,9 +49,6 @@ sourceSets {
 configurations {
   functionalTestImplementation.extendsFrom testImplementation
   functionalTestRuntimeOnly.extendsFrom runtimeOnly
-
-  integrationTestImplementation.extendsFrom testImplementation
-  integrationTestRuntimeOnly.extendsFrom runtimeOnly
 
   smokeTestImplementation.extendsFrom testImplementation
   smokeTestRuntimeOnly.extendsFrom runtimeOnly
@@ -109,14 +97,6 @@ task functional(type: Test) {
   classpath = sourceSets.functionalTest.runtimeClasspath
 }
 
-task integration(type: Test) {
-  description = "Runs integration tests"
-  group = "Verification"
-  testClassesDirs = sourceSets.integrationTest.output.classesDirs
-  classpath = sourceSets.integrationTest.runtimeClasspath
-  failFast = true
-}
-
 task smoke(type: Test) {
   description = "Runs Smoke Tests"
   testClassesDirs = sourceSets.smokeTest.output.classesDirs
@@ -131,7 +111,7 @@ checkstyle {
 
 pmd {
   toolVersion = "6.43.0"
-  sourceSets = [sourceSets.main, sourceSets.test, sourceSets.functionalTest, sourceSets.integrationTest, sourceSets.smokeTest]
+  sourceSets = [sourceSets.main, sourceSets.test, sourceSets.functionalTest, sourceSets.smokeTest]
   reportsDir = file("$project.buildDir/reports/pmd")
   // https://github.com/pmd/pmd/issues/876
   ruleSets = []
@@ -215,18 +195,18 @@ dependencies {
   implementation group: 'com.github.hmcts.java-logging', name: 'logging-appinsights', version: reformLoggingVersion
   implementation group: 'net.logstash.logback', name: 'logstash-logback-encoder', version: '7.3'
 
-  implementation group: 'com.azure.spring', name: 'spring-cloud-azure-starter-active-directory', version: '4.7.0'
+  implementation group: 'com.azure.spring', name: 'spring-cloud-azure-starter-active-directory', version: '4.8.0'
   implementation group: 'org.postgresql', name: 'postgresql', version: '42.6.0'
-  implementation group: 'com.github.hmcts', name: 'pip-data-models', version: '2.0.4', {
+  implementation group: 'com.github.hmcts', name: 'pip-data-models', version: '2.0.5', {
     exclude group: 'org.springframework.boot', module: 'spring-boot-starter-data-jpa'
   }
   implementation group: 'com.squareup.okhttp3', name: 'okhttp', version: '4.11.0'
   implementation group: 'com.vladmihalcea', name: 'hibernate-types-52', version: '2.21.1'
 
-  implementation group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-jsr310', version: '2.15.0'
+  implementation group: 'com.fasterxml.jackson.datatype', name: 'jackson-datatype-jsr310', version: '2.15.2'
 
   // Include Flyway for database migrations
-  implementation group: 'org.flywaydb', name: 'flyway-core', version: '9.17.0'
+  implementation group: 'org.flywaydb', name: 'flyway-core', version: '9.19.1'
 
   // Force upgrade snakeyaml version for CVE-2022-38752
   implementation( group: 'org.yaml', name: 'snakeyaml').version {
@@ -246,7 +226,7 @@ dependencies {
 
   testImplementation group: 'org.apiguardian', name: 'apiguardian-api', version: '1.1.2'
 
-  testImplementation group: 'io.zonky.test', name: 'embedded-database-spring-test', version: '2.2.0'
+  testImplementation group: 'io.zonky.test', name: 'embedded-database-spring-test', version: '2.3.0'
   testImplementation group: 'io.github.hakky54', name: 'logcaptor', version: '2.9.0'
   testImplementation group: 'com.squareup.okhttp3', name: 'mockwebserver', version: '4.11.0'
   testImplementation group: 'org.springframework.security', name: 'spring-security-test'
@@ -266,7 +246,7 @@ rootProject.tasks.named("processSmokeTestResources") {
 }
 
 //To use multiple application.yamls (for db mocking)
-rootProject.tasks.named("processIntegrationTestResources") {
+rootProject.tasks.named("processFunctionalTestResources") {
   duplicatesStrategy = 'include'
 }
