+++ charts/pip-subscription-management/Chart.yaml
@@ -3,10 +3,10 @@ appVersion: "1.0"
 description: A Helm chart for pip-subscription-management App
 name: pip-subscription-management
 home: https://github.com/hmcts/pip-subscription-management
-version: 0.0.24
+version: 0.0.25
 maintainers:
   - name: HMCTS PIP Team
 dependencies:
   - name: java
-    version: 4.1.0
+    version: 4.1.2
     repository: 'https://hmctspublic.azurecr.io/helm/v1/repo/'
