+++ config/owasp/suppressions.xml
@@ -2,9 +2,9 @@
 <suppressions xmlns="https://jeremylong.github.io/DependencyCheck/dependency-suppression.1.3.xsd">
   <suppress>
     <notes><![CDATA[
-   file name: spring-cloud-azure-starter-active-directory-4.7.0.jar
+   file name: spring-cloud-azure-starter-active-directory-4.8.0.jar
    ]]></notes>
-    <packageUrl regex="true">^pkg:maven/com.azure.spring/spring-cloud-azure-starter-active-directory@4.7.0</packageUrl>
+    <packageUrl regex="true">^pkg:maven/com.azure.spring/spring-cloud-azure-starter-active-directory@4.8.0</packageUrl>
     <cve>CVE-2021-42306</cve>
   </suppress>
   <suppress>
