+++ src/functionalTest/java/uk/gov/hmcts/reform/pip/subscription/management/config/WebClientConfigurationTest.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.web.reactive.function.client.WebClient;
 
 @Configuration
-@Profile("integration")
+@Profile("functional")
 public class WebClientConfigurationTest {
 
     @Bean
