+++ src/functionalTest/java/uk/gov/hmcts/reform/pip/subscription/management/controllers/SubscriptionControllerTests.java
@@ -41,7 +41,6 @@ import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -50,7 +49,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @SpringBootTest(classes = {Application.class, WebClientConfigurationTest.class},
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @AutoConfigureMockMvc
-@ActiveProfiles("integration")
+@ActiveProfiles("functional")
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
 @WithMockUser(username = "admin", authorities = {"APPROLE_api.request.admin"})
 @AutoConfigureEmbeddedDatabase(type = AutoConfigureEmbeddedDatabase.DatabaseType.POSTGRES)
@@ -82,6 +81,7 @@ class SubscriptionControllerTests {
     private static final String VALIDATION_NO_SUBSCRIPTIONS = "User has unknown subscriptions";
     public static final String VALIDATION_ONE_CASE_LOCATION = "Location subscription list does not contain 1 case";
     public static final String VALIDATION_DATE_ADDED = "Date added does not match the expected date added";
+    public static final String VALIDATION_UUID = "UUID should not be null";
     private static final String FORBIDDEN_STATUS_CODE = "Status code does not match forbidden";
     private static final String NOT_FOUND_STATUS_CODE = "Status code does not match not found";
     private static final String RESPONSE_MATCH = "Response should match";
@@ -237,9 +237,9 @@ class SubscriptionControllerTests {
             returnedSubscription.getUserId(),
             VALIDATION_USER_ID
         );
-        assertNotEquals(
-            returnedSubscription.getId(), 0L, "id should not equal zero"
-        );
+
+        assertNotNull(returnedSubscription.getId(), VALIDATION_UUID);
+
         assertEquals(CASE_NAME, returnedSubscription.getCaseName(),
                      VALIDATION_CASE_NAME
         );
@@ -305,22 +305,21 @@ class SubscriptionControllerTests {
             returnedSubscription2.getUserId(),
             VALIDATION_USER_ID
         );
-        assertNotEquals(
-            returnedSubscription2.getId(), 0L, "id should not equal zero"
-        );
-        assertEquals(CASE_NAME, returnedSubscription.getCaseName(),
+        assertNotNull(returnedSubscription2.getId(), VALIDATION_UUID);
+
+        assertEquals(CASE_NAME, returnedSubscription2.getCaseName(),
                      VALIDATION_CASE_NAME
         );
-        assertEquals(CASE_ID, returnedSubscription.getCaseNumber(),
+        assertEquals(CASE_ID, returnedSubscription2.getCaseNumber(),
                      VALIDATION_CASE_NUMBER
         );
-        assertEquals(CASE_URN, returnedSubscription.getUrn(),
+        assertEquals(CASE_URN, returnedSubscription2.getUrn(),
                      VALIDATION_CASE_URN
         );
-        assertEquals(PARTY_NAMES, returnedSubscription.getPartyNames(),
+        assertEquals(PARTY_NAMES, returnedSubscription2.getPartyNames(),
                      VALIDATION_PARTY_NAMES
         );
-        assertEquals(LOCATION_NAME_1, returnedSubscription.getLocationName(),
+        assertEquals(LOCATION_NAME_1, returnedSubscription2.getLocationName(),
                      VALIDATION_LOCATION_NAME
         );
