+++ src/functionalTest/resources/mock/artefact.json
